# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: FinDoc Analyzer API
  description: |
    API for FinDoc Analyzer application that processes financial documents, 
    extracts securities data, and provides portfolio analysis.
  contact:
    name: FinDoc Support
    url: https://findoc.example.com/support
    email: support@findoc.example.com
  version: 1.0.0
servers:
  - url: https://api.findoc.example.com/v1
    description: Production server
  - url: https://api-staging.findoc.example.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /documents:
    get:
      tags:
        - documents
      summary: Get documents
      description: Retrieve a list of documents for the authenticated user
      operationId: getDocuments
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [uploadDate, fileName, processingStatus]
            default: uploadDate
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: filter
          in: query
          description: Filter by processing status
          required: false
          schema:
            type: string
            enum: [all, pending, processing, completed, failed]
            default: all
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /documents/{documentId}:
    get:
      tags:
        - documents
      summary: Get document
      description: Retrieve document details by ID
      operationId: getDocument
      parameters:
        - name: documentId
          in: path
          description: Document ID
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  document:
                    $ref: '#/components/schemas/Document'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - documents
      summary: Delete document
      description: Delete a document by ID
      operationId: deleteDocument
      parameters:
        - name: documentId
          in: path
          description: Document ID
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Document deleted successfully
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /documents/process:
    post:
      tags:
        - documents
        - processor
      summary: Process document
      description: Start processing a document
      operationId: processDocument
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - documentId
              properties:
                documentId:
                  type: string
                  example: doc-123
                options:
                  type: object
                  properties:
                    useOcr:
                      type: boolean
                      description: Whether to use OCR for processing
                      default: false
                    extractTables:
                      type: boolean
                      description: Whether to extract tables
                      default: true
                    extractSecurities:
                      type: boolean
                      description: Whether to extract securities
                      default: true
                    processingTier:
                      type: string
                      enum: [basic, standard, comprehensive]
                      description: Processing tier level
                      default: standard
      responses:
        '200':
          description: Processing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  processingId:
                    type: string
                    example: proc-456
                  documentId:
                    type: string
                    example: doc-123
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          example: doc-123
        fileName:
          type: string
          example: portfolio-statement.pdf
        fileType:
          type: string
          example: pdf
        fileSize:
          type: integer
          format: int64
          example: 1048576
        uploadDate:
          type: string
          format: date-time
          example: '2023-01-15T12:00:00Z'
        processingStatus:
          type: string
          enum: [pending, processing, completed, failed]
          example: completed
        processingDate:
          type: string
          format: date-time
          example: '2023-01-15T12:05:00Z'
        description:
          type: string
          example: Q1 2023 Portfolio Statement
        tags:
          type: array
          items:
            type: string
          example: [portfolio, q1, 2023]
        metadata:
          type: object
          properties:
            pageCount:
              type: integer
              example: 5
            language:
              type: string
              example: en
            securitiesCount:
              type: integer
              example: 8
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 54
        totalPages:
          type: integer
          example: 3
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Document not found
        code:
          type: string
          example: DOCUMENT_NOT_FOUND
        status:
          type: integer
          example: 404